+++ src/main/java/com/example/DUT_Parking/services/RegisterService.java
@@ -1,6 +1,7 @@
 package com.example.DUT_Parking.services;
 
 import com.example.DUT_Parking.DTO.RegisterRequest;
+import com.example.DUT_Parking.exception_handling.AppException;
 import com.example.DUT_Parking.respond.GetRegisteredUsers;
 
 import java.util.List;
@@ -9,6 +10,6 @@ public interface RegisterService {
 
     boolean register(RegisterRequest registerRequest);
     List<GetRegisteredUsers> getAllUsers();
-    void delete(int id);
+    void delete(Long id);
     GetRegisteredUsers search(String email);
 }
